+++ src/bigdict/__init__.py
@@ -1,7 +1,7 @@
 '''
 Bigdict
 '''
-__version__ = "0.2.1"
+__version__ = "0.2.2b1"
 
 from ._bigdict import Bigdict